Add getUserProgress function
This function fetches the users progress data from the database. It first retrieves the userId using auth() from @clerk/nextjs. If the userId is not available, it returns null. Otherwise, it queries the userProgress table to find the user's progress based on the userId. The result includes the activeCourse information.